Map preload (WIP).
Map tiles near the event location begin to preload when EventActivity is opened. When the user leaves, the preloading stops. Before preload there's a check, what map tiles are already in
the disk cache. This takes about 1 ms for every tile, which can be slow when there are lot of tiles.